Fix Amount conversion errors Remove superflous PriceDto class

+++ src/SwedbankPay.Sdk.Infrastructure/PaymentInstruments/PriceListDto.cs
@@ -3,7 +3,7 @@
     internal class PriceListDto
     {
         public string Type { get; set; }
-        public int Amount { get; set; }
-        public int VatAmount { get; set; }
+        public long Amount { get; set; }
+        public long VatAmount { get; set; }
     }
 }
